requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.4.0
apturl==0.5.2
async-timeout==4.0.3
attrs==23.2.0
auth0-python==4.4.2
bcrypt==3.2.0
billiard==4.2.0
blinker==1.7.0
boto3==1.28.59
botocore==1.31.85
Brlapi==0.8.3
bytecode==0.15.1
cached-property==1.5.2
cachetools==5.0.0
cattrs==23.2.3
celery==5.3.4
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
cliapp==1.20180812.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
coverage==7.3.2
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.2.18
ddsketch==2.0.4
ddtrace==1.19.0
defer==1.0.6
dell-recovery==0.0.0
Deprecated==1.2.14
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.6.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docusign-esign==3.24.0
duplicity==0.8.21
ecdsa==0.18.0
email_validator==2.2.0
envier==0.5.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
fasteners==0.14.1
flasgger==0.9.7.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
Flask-Log-Request-ID==0.10.1
Flask-Migrate==4.0.5
Flask-Pydantic==0.11.0
flask-restx==1.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
future==0.18.2
google-api-core==2.17.1
google-api-python-client==2.101.0
google-auth==2.28.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
googleapis-common-protos==1.62.0
greenlet==1.1.2
gunicorn==21.2.0
gyp==0.1
h11==0.14.0
httpagentparser==1.9.5
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
idna==3.3
importlib-metadata==6.11.0
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.19.1
jsonschema-specifications==2023.12.1
jwcrypto==1.5.1
keyring==23.5.0
kombu==5.3.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==5.1.0
macaroonbakery==1.3.1
mailchimp-transactional==1.0.50
Mako==1.1.3
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mistune==3.0.2
monotonic==1.6
more-itertools==8.10.0
msoffcrypto-tool==5.1.1
mt-940==4.30.0
multidict==6.0.5
netifaces==0.11.0
nose==1.3.7
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.2
opentelemetry-api==1.22.0
orjson==3.10.6
packaging==23.2
pandas==2.1.1
paramiko==3.4.0
passlib==1.7.4
pdfkit==1.0.0
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.0
progressbar==2.5
prompt-toolkit==3.0.43
protobuf==4.25.3
psycopg2==2.9.9
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycparser==2.21
pycryptodome==3.19.1
pycups==2.0.1
pydantic==2.6.1
pydantic_core==2.16.2
PyDrive==1.3.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
pymongo==4.8.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pysftp==0.2.9
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-gnupg==0.5.1
python-intercom==3.1.0
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.9
pytz==2023.3.post1
pyxdg==0.27
PyYAML==5.4.1
redis==5.0.1
referencing==0.33.0
reportlab==3.6.8
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.8
s3transfer==0.7.0
SecretStorage==3.3.1
semantic-version==2.10.0
sentry-sdk==1.31.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sos==4.5.6
SQLAlchemy==2.0.21
sqlalchemy-json==0.7.0
SQLAlchemy-Utils==0.41.1
starlette==0.37.2
structlog==23.1.0
systemd-python==234
texttable==1.6.4
ttystatus==0.38
typer==0.12.3
typing_extensions==4.9.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.10.0
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==2.0.7
usb-creator==0.3.7
uvicorn==0.30.1
uvloop==0.19.0
vboxapi==1.0
vine==5.1.0
wadllib==1.3.6
watchfiles==0.22.0
wcwidth==0.2.13
websocket-client==1.2.3
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
xdg==5
xkit==0.0.0
xlrd==2.0.1
XlsxWriter==3.1.6
xmltodict==0.13.0
yarl==1.9.4
zeep==4.2.1
zipp==1.0.0
stripe==10.3.0
pytest==8.2.2
